move inspect into internal from helpers
