+++ README.md
@@ -34,8 +34,8 @@ const { Maybe, compose, curry, map } = crocks
 // I feel ya, try this:
 
 import Maybe from 'crocks/crocks/Maybe'
-import compose from 'crocks/funcs/compose'
-import curry from 'crocks/funcs/curry'
+import compose from 'crocks/helpers/compose'
+import curry from 'crocks/helpers/curry'
 import map from 'crocks/pointfree/map'
 
 // you can of course do the same with require statements:
@@ -52,7 +52,7 @@ There are (5) classifications of "things" included in this library:
 
 * Combinators (`crocks/combinators`): A collection of functions that are used for working with other functions. These do things like compose (2) functions together, or flip arguments on a function. They typically either take a function, return a function or a bit a both. These are considered the glue that holds the mighty house of `Crocks` together and a valuable aid in writing reusable code.
 
-* Helper Functions (`crocks/funcs`): All other support functions that are either convenient versions of combinators or not even combinators at all cover this group.
+* Helper Functions (`crocks/helpers`): All other support functions that are either convenient versions of combinators or not even combinators at all cover this group.
 
 * Point-free Functions (`crocks/pointfree`): Wanna use these ADTs in a way that you never have to reference the actual data being worked on? Well here is where you will find all of these functions to do that. For every algebra available on both the `Crocks` and `Monoids` there is a function here.
 
@@ -101,7 +101,7 @@ All `Monoids` provide `empty` and `type` function on their Constructors as well 
 Seems really silly, but is quite useful for a lot of things. It takes a function and a value and then returns the result of that function with the argument applied.
 
 #### `composeB : (b -> c) -> (a -> b) -> a -> c`
-Provides a means to describe a composition between two functions. it takes two functions and an value. Given `composeB(f, g)`, which is read `f` after `g`, it will return a function that will take value `a` and apply it to `g`, passing the result as an argument to `f`, and will finally return the result of `f`. (This allows only two functions, if you want to avoid things like: `composeB(composeB(f, g), composeB(h, i))` then check out `crocks/funcs/compose`.)
+Provides a means to describe a composition between two functions. it takes two functions and an value. Given `composeB(f, g)`, which is read `f` after `g`, it will return a function that will take value `a` and apply it to `g`, passing the result as an argument to `f`, and will finally return the result of `f`. (This allows only two functions, if you want to avoid things like: `composeB(composeB(f, g), composeB(h, i))` then check out `crocks/helpers/compose`.)
 
 #### `constant : a -> b -> a`
 This is a very handy dandy function, used a lot. Pass it any value and it will give you back a function that will return that same value no matter what you pass it.
