+++ crocks/Arrow.js
@@ -3,8 +3,8 @@
 
 const isFunction = require('../internal/isFunction')
 const isType = require('../internal/isType')
+const _inspect = require('../internal/inspect')
 
-const _inspect = require('../helpers/inspect')
 const compose = require('../helpers/compose')
 
 const identity = require('../combinators/identity')
