+++ crocks/Either.js
@@ -5,13 +5,12 @@ const isApplicative = require('../internal/isApplicative')
 const isFunction = require('../internal/isFunction')
 const isType = require('../internal/isType')
 
+const _inspect = require('../internal/inspect')
 const defineUnion = require('../internal/defineUnion')
 
 const constant = require('../combinators/constant')
 const composeB = require('../combinators/composeB')
 const identity = require('../combinators/identity')
-
-const _inspect = require('../helpers/inspect')
 
 const _either = defineUnion({ Left: [ 'a' ], Right: [ 'b' ] })
