+++ crocks/List.js
@@ -6,9 +6,10 @@ const isArray = require('../internal/isArray')
 const isType = require('../internal/isType')
 const isApplicative = require('../internal/isApplicative')
 
+const _inspect = require('../internal/inspect')
+
 const constant = require('../combinators/constant')
 
-const _inspect = require('../helpers/inspect')
 const _concat = require('../pointfree/concat')
 
 const Maybe = require('./Maybe')
