+++ monoids/Any.js
@@ -5,7 +5,7 @@ const isType = require('../internal/isType')
 const isFunction = require('../internal/isFunction')
 const isNil = require('../internal/isNil')
 
-const _inspect = require('../helpers/inspect')
+const _inspect = require('../internal/inspect')
 
 const constant = require('../combinators/constant')
