+++ monoids/Sum.js
@@ -6,7 +6,7 @@ const isNumber = require('../internal/isNumber')
 const isFunction = require('../internal/isFunction')
 const isNil = require('../internal/isNil')
 
-const _inspect = require('../helpers/inspect')
+const _inspect = require('../internal/inspect')
 
 const constant = require('../combinators/constant')
