+++ test/LastMonoid.js
@@ -1,6 +1,7 @@
 const identity = require('../combinators/identity')
 const constant = require('../combinators/constant')
-const _inspect = require('../helpers/inspect')
+
+const _inspect = require('../internal/inspect')
 
 const _type = constant('Last')
